+++ backend/backend.go
@@ -24,13 +24,24 @@ type Key struct {
 
 // Profile describes a single profile in a .kiya configuration
 type Profile struct {
-	Backend     string
-	Label       string
-	ProjectID   string
-	Location    string
-	Keyring     string
-	CryptoKey   string
-	Bucket      string
-	VaultUrl    string
-	SecretRunes []rune
+	Backend string
+	// General
+	Label               string
+	Location            string
+	SecretRunes         []rune
+	AutoCopyEnabled     bool `json:"autoCopyEnabled"`     // if true then the secret of a single list result will be copied to clipboard
+	PromptForSecretLine bool `json:"promptForSecretLine"` // if true then you must enter a number to run the command on that line
+
+	// GCP
+	ProjectID string
+	Keyring   string
+	CryptoKey string
+	Bucket    string
+
+	// AWS
+	AWSProfile string `json:"awsprofile"`
+
+	// Vault
+	VaultUrl       string
+	VaultMountPath string
 }