+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Changes
 
+### v1.14.0
+
+- Add PromptForCommandLine field to profile.
+
+### v1.13.0
+
+- Add AutoCopyEnabled field to profile.
+
 ### v1.12.5
 
 - Security upgrades from Dependabot